refactor(lungo): replace useElementSize with useElementHeight and elementSize type
