+++ coffeeAGNTCY/coffee_agents/lungo/frontend/src/hooks/layout/elementSize.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright AGNTCY Contributors (https://github.com/agntcy)
+ * SPDX-License-Identifier: Apache-2.0
+ **/
+
+export interface ElementSize {
+  width: number
+  height: number
+}
