+++ coffeeAGNTCY/coffee_agents/lungo/frontend/src/hooks/layout/index.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  **/
 
+export type { ElementSize } from "./elementSize"
 export { useElementWidth } from "./useElementWidth"
-export { useElementSize } from "./useElementSize"
-export type { ElementSize } from "./useElementSize"
+export { useElementHeight } from "./useElementHeight"
 export { useElementRect } from "./useElementRect"
 export { useIsBelowSmBreakpoint } from "./useIsBelowSmBreakpoint"
